Update optimalmean.py
the procedure turned out to be very sensitive to the array being distributed into bags with exactly the same element size.

fixed the partitioning procedure such that the confidence interval is automatically increased to the next value where even partitioning is possible. 

This increases the stability of the algorithm much if it is used with asymmetric distributions. 

Unfortunately, the mean of the exponential distribution is still often not estimated as good as the ordinary sample mean.